Split examples into main programs and fmha runners, build instances separately
